+++ libopenage/renderer/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_sources(libopenage
 	color.cpp
 	context.cpp
+	program.cpp
 	renderer.cpp
 	shader.cpp
 	task.cpp
