+++ libopenage/renderer/shader.h
@@ -84,60 +84,6 @@ public:
 	virtual ~ShaderSourceCode() {};
 };
 
-
-/**
- * Create this to assemble shaders to a usable program.
- * Register it at the renderer to use it.
- *
- * A program is a graphics card kernel, consisting of multiple
- * shaders used for the rendering stages.
- */
-class ProgramSource {
-public:
-	ProgramSource();
-	ProgramSource(const std::vector<const ShaderSource *> &shaders);
-	virtual ~ProgramSource() {};
-
-	/**
-	 * Attach the given shader to this program.
-	 */
-	void attach_shader(const ShaderSource &shader);
-
-	/**
-	 * Return the list of all assigned shaders to this program.
-	 */
-	const std::vector<const ShaderSource *> &get_shaders() const;
-
-protected:
-	/**
-	 * All attached shaders to this program.
-	 */
-	std::vector<const ShaderSource *> shaders;
-};
-
-
-/**
- * A usable handle, aquired by registering the sources of the program
- * to the renderer and getting back an object of this class.
- */
-class RawProgram {
-protected:
-	RawProgram();
-
-public:
-	virtual ~RawProgram() {};
-
-	/**
-	 * Use this program now on the GPU.
-	 */
-	virtual void use() = 0;
-
-	/**
-	 * Dump vertex attribute variables.
-	 */
-	virtual void dump_attributes() = 0;
-};
-
 }} // openage::renderer
 
 #endif
